requirements.txt
opentelemetry-api~=1.23.0
opentelemetry-sdk~=1.23.0
opentelemetry-exporter-otlp~=1.23.0
opentelemetry-instrumentation~=0.44b0
